moved the last repo command to its own file.
